Style(#6): BestView 추가
